requirements.txt
streamlit==1.28.1
streamlit-calendar==1.4.0
pandas==2.1.3
bcrypt==4.1.2
bleach==6.1.0
psutil==5.9.6
plotly==5.17.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.3.4
python-dotenv==1.0.0
alembic==1.13.1
requests==2.31.0
phonenumbers==8.13.26
openpyxl==3.1.2
reportlab==4.0.7
pywhatkit==5.4
selenium==4.15.2
webdriver-manager==4.0.1
qrcode==7.4.2
pillow==10.1.0